+++ include/mapping/front_end/front_end.hpp
@@ -21,22 +21,22 @@ namespace avp_mapping {
 
 class FrontEnd {
 public:
-FrontEnd(float key_frame_dist, int local_frame_num);
+FrontEnd(double key_frame_dist, int local_frame_num);
 FrontEnd() = default;
 
 struct Frame{
-  Eigen::Matrix4f pose = Eigen::Matrix4f::Identity();
+  Eigen::Matrix4d pose = Eigen::Matrix4d::Identity();
   CloudData cloud_data;
 };
 
-bool setInitPose(const Eigen::Matrix4f init_pose);
-bool update(const CloudData& cloud_data, Eigen::Matrix4f& updated_pose);
+bool setInitPose(const Eigen::Matrix4d init_pose);
+bool update(const CloudData& cloud_data, Eigen::Matrix4d& updated_pose);
 
 
 private:
 
 bool initRegistration(std::shared_ptr<RegistrationInterface>& registration_ptr, const std::string& type);
-bool initFilter(std::shared_ptr<CloudFilterInterface>& filter_ptr, const std::string& type, float leaf_size);
+bool initFilter(std::shared_ptr<CloudFilterInterface>& filter_ptr, const std::string& type, double leaf_size);
 bool udpateWithNewKeyFrame(const Frame& new_key_frame);
 
 std::shared_ptr<CloudFilterInterface> frame_filter_ptr_;
@@ -48,9 +48,9 @@ std::deque<Frame> local_map_frames_;
 CloudData::CLOUD_PTR local_map_ptr_;
 Frame current_frame_;
 
-Eigen::Matrix4f init_pose_ = Eigen::Matrix4f::Identity();
+Eigen::Matrix4d init_pose_ = Eigen::Matrix4d::Identity();
 
-float key_frame_distance_ = 2.0;
+double key_frame_distance_ = 2.0;
 int local_frame_num_ = 20;
 
 };
